+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/feat/SetupGradlePublishPlugin.kt
@@ -1,14 +1,14 @@
-package fluxo.conf.impl.kotlin
+package fluxo.conf.feat
 
-import fluxo.conf.FluxoKmpConfContext
 import fluxo.conf.data.BuildConstants.GRADLE_PLUGIN_PUBLISH_PLUGIN_ALIAS
 import fluxo.conf.data.BuildConstants.GRADLE_PLUGIN_PUBLISH_PLUGIN_ID
 import fluxo.conf.data.BuildConstants.GRADLE_PLUGIN_PUBLISH_PLUGIN_VERSION
 import fluxo.conf.deps.loadAndApplyPluginIfNotApplied
+import fluxo.conf.dsl.impl.FluxoConfigurationExtensionImpl
 import org.gradle.api.Project
 
-internal fun Project.setupGradlePublishPlugin(ctx: FluxoKmpConfContext) {
-    ctx.loadAndApplyPluginIfNotApplied(
+internal fun Project.setupGradlePublishPlugin(conf: FluxoConfigurationExtensionImpl) {
+    conf.ctx.loadAndApplyPluginIfNotApplied(
         id = GRADLE_PLUGIN_PUBLISH_PLUGIN_ID,
         version = GRADLE_PLUGIN_PUBLISH_PLUGIN_VERSION,
         catalogPluginId = GRADLE_PLUGIN_PUBLISH_PLUGIN_ALIAS,
@@ -19,6 +19,8 @@ internal fun Project.setupGradlePublishPlugin(ctx: FluxoKmpConfContext) {
         lookupClassName = false,
 //        className = GRADLE_PLUGIN_PUBLISH_CLASS_NAME,
     )
+
+    setupValidatePluginTasks(conf)
 }
 
 @Suppress("unused")