+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Added
 - auto downgrade Kotlin API version when it's greater than the language version (and warn about it).
+- set up `validatePlugins` task for Gradle plugins. **BREAKING CHANGE!**
 
 ### Changed
 - prepare for Kotlin 2.0.